Refactor: Review 리팩토링
