+++ src/main/java/org/springeel/oneclickrecipe/domain/recipeprocess/dto/service/response/RecipeProcessReadResponseDto.java
@@ -1,6 +1,7 @@
 package org.springeel.oneclickrecipe.domain.recipeprocess.dto.service.response;
 
 public record RecipeProcessReadResponseDto(
+    Long id,
     Byte sequence,
     String description,
     Short time,
