+++ src/main/java/org/springeel/oneclickrecipe/domain/review/dto/service/request/ReviewCreateServiceRequestDto.java
@@ -1,4 +1,4 @@
-package org.springeel.oneclickrecipe.domain.review.dto.service;
+package org.springeel.oneclickrecipe.domain.review.dto.service.request;
 
 public record ReviewCreateServiceRequestDto(
     String content,
