+++ src/main/java/org/springeel/oneclickrecipe/domain/review/dto/service/response/ReviewReadResponseDto.java
@@ -1,9 +1,10 @@
-package org.springeel.oneclickrecipe.domain.review.dto.service;
+package org.springeel.oneclickrecipe.domain.review.dto.service.response;
 
 import lombok.Builder;
 
 @Builder
 public record ReviewReadResponseDto(
+    Long id,
     String content,
     Byte star
 ) {
