+++ src/main/java/org/springeel/oneclickrecipe/domain/review/mapper/dto/ReviewDtoMapper.java
@@ -1,13 +1,14 @@
 package org.springeel.oneclickrecipe.domain.review.mapper.dto;
 
+import static org.mapstruct.MappingConstants.ComponentModel.SPRING;
+
 import org.mapstruct.Mapper;
-import org.mapstruct.MappingConstants;
 import org.springeel.oneclickrecipe.domain.review.dto.controller.ReviewCreateControllerRequestDto;
 import org.springeel.oneclickrecipe.domain.review.dto.controller.ReviewUpdateControllerRequestDto;
-import org.springeel.oneclickrecipe.domain.review.dto.service.ReviewCreateServiceRequestDto;
-import org.springeel.oneclickrecipe.domain.review.dto.service.ReviewUpdateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.review.dto.service.request.ReviewCreateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.review.dto.service.request.ReviewUpdateServiceRequestDto;
 
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
+@Mapper(componentModel = SPRING)
 public interface ReviewDtoMapper {
 
     ReviewCreateServiceRequestDto toReviewCreateServiceRequestDto(
