+++ src/main/java/org/springeel/oneclickrecipe/domain/review/mapper/entity/ReviewEntityMapper.java
@@ -1,19 +1,20 @@
 package org.springeel.oneclickrecipe.domain.review.mapper.entity;
 
+import static org.mapstruct.MappingConstants.ComponentModel.SPRING;
+
 import java.util.List;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.mapstruct.MappingConstants;
 import org.springeel.oneclickrecipe.domain.recipe.entity.Recipe;
-import org.springeel.oneclickrecipe.domain.review.dto.service.ReviewCreateServiceRequestDto;
-import org.springeel.oneclickrecipe.domain.review.dto.service.ReviewReadResponseDto;
+import org.springeel.oneclickrecipe.domain.review.dto.service.request.ReviewCreateServiceRequestDto;
+import org.springeel.oneclickrecipe.domain.review.dto.service.response.ReviewReadResponseDto;
 import org.springeel.oneclickrecipe.domain.review.entity.Review;
 import org.springeel.oneclickrecipe.domain.user.entity.User;
 
-@Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
+@Mapper(componentModel = SPRING)
 public interface ReviewEntityMapper {
 
-    List<ReviewReadResponseDto> toReviewResponseDtos(List<Review> review);
+    List<ReviewReadResponseDto> toReviewReadResponseDtos(List<Review> reviews);
 
     ReviewReadResponseDto toReviewReadResponseDto(Review review);
