+++ src/main/java/org/springeel/oneclickrecipe/domain/review/repository/ReviewRepository.java
@@ -10,5 +10,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
 
     Optional<Review> findByIdAndUser(Long recipeId, User user);
 
-    List<Review> findByRecipeId(Long recipeId);
+    List<Review> findAllByRecipeId(Long recipeId);
 }
